refactor: ai, aladin 패키지 infrastructure 패키지로 이동 #131
