--- src/main/java/com/bookripple/api/domain/aladin/dto/AladinSearchResDto.java
@@ -1,24 +1,0 @@
-package com.bookripple.api.domain.aladin.dto;
-
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-public class AladinSearchResDto {
-    private Integer totalResults;
-    private Integer startIndex;
-    private Integer itemsPerPage;
-    private List<Item> item;
-
-    @Getter
-    public static class Item {
-        private Long itemId;        // 알라딘 itemId
-        private String title;
-        private String author;
-        private String publisher;
-        private String pubDate;     // "yyyy-MM-dd"
-        private String cover;
-        private String isbn13;
-    }
-}
