+++ src/main/java/com/bookripple/api/infrastructure/aladin/config/AladinConfig.java
@@ -1,4 +1,4 @@
-package com.bookripple.api.domain.aladin.config;
+package com.bookripple.api.infrastructure.aladin.config;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
@@ -6,4 +6,5 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 @EnableConfigurationProperties(AladinProperties.class)
 public class AladinConfig {
+
 }
